Fix g-ir-introspection
When generating gobject-introspection, the python program g-ir-introspection is
patched to find the python modules and other elements at $CRAFT_STAGE. Unfortunately,
the environment name is inserted "as is" in the strings, instead of using os.environ
to get the value.

This makes fail ATK, because when it tries to execute g-ir-scanner, it can't find
one of the python modules that it needs to run.

This patch fixes this, by using os.environ.get('CRAFT_STAGE') instead.