Update Pango version to fix Gtk4 build
The current Pango version is 1.48.0, but in order to buid Gtk4,
a precompiled binary, 'pango-view' must be run. This binary is
compiled against Pango 1.50.6, which is the version currently
available in Ubuntu 22.04, and uses an exported function,
'pango_glyph_string_index_to_x_full()', that was added precisely
in Pango 1.50 (https://docs.gtk.org/Pango/struct.GlyphString.html).

This patch updates Pango version to 1.50.6 to fix this.