Adapt XXXmm versions to ensure useful building
When building gtkmm 3.24, if a too much recent version of glibmm
is detected, the builder will not use it and will download a
suitable old version. The same happens with cairomm and pangomm.

The net result is that we are compiling and adding code that
won't be used or needed, and that only takes space.

This MR adjusts the versions of all these parts to the last
valid version. Also adds some extra comments to avoid to
repeat this mistake in the future.