DT-488 Add support for AMDGPU_ASIC_ID_TABLE_PATH
Every time a GTK-based program is launched in a computer with
an AMD-based GPU, the error

    /usr/share/libdrm/amdgpu.ids: No such file or directory

is shown in the terminal. The reason is that libdrm expects
that file to be in that precise location. Although the file
is now available in Gnome-42-2204, it isn’t accessible
directly, but each snap has to manually add a LAYOUT statement
to link that file from
$SNAP/gnome-contents/usr/share/libdrm to the root.

A possible fix would be to set that LAYOUT statement in
snapcraft, in the "gnome extension" code, but that means that
any snap that wants to include its own libdrm version won't be
able to put its own version of the file.

This patch fixes this by adding a new environment variable to
libdrm: AMDGPU_ASIC_ID_TABLE_PATH. This variable holds the
full path to the file, and it is used only when the file isn't
found in the expected location. Thus any snap that wants to
include its own version of libdrm can also use a LAYOUT
statement to place its own file at /usr/share/libdrm, while the
Gnome Extension snap can use the environment variable to point
to its file.

This patch requires https://github.com/snapcore/snapcraft-desktop-integration/pull/5
to work.