Include the fonts configuration from a default desktop installation
We need at least the URW fonts configuration for correct PDF rendering as
reported against firefox on https://github.com/ubuntu/gnome-sdk/issues/49

After discussion we decided to include the configurations provided in a
default Ubuntu desktop installation for 42. It's not ideal but should
help with consistency with non snaps apps and be an improvement.
Going forward we should work with fontconfig upstream on a solution
which allows use to import the system configuration instead.