Use local gtk binaries
The "caches" part installs gtk-update-icon-cache, libglib2.0-bin
and shared-mime-info .deb packages to use some of their binaries.
Unfortunately, the $PATH puts before the built binaries instead
of the ones in the system, but the libraries are from the system.
This can result in a failure if a new version being compiled
requires a new call in the libraries.

This patch fixes it.