Remove invalid pkgconfig files and broken links
Some soft links to libraries that are already in Core 24, and
thus, don't needed, are kept in the SDK.

Also, two pkgconfig files are also kept, pointing to the SDK
library folder. But since those two libraries (libcrypt and
libpciaccess) are already in Core24, and the libraries
themselves are deleted, this breaks any build that tries to
use any of those.

Removing the broken soft links and the pkgconfig files fixes
this. With this patch, installing the development .deb packages
using the "build-packages" statement in the snapcraft.yaml
ensures that the requires build files are available and that no
other pkgconfig file will override the correct ones.