Allow building with a local SDK snap
The RUNTIME snap is derived from the corresponding SDK snap, but
currently, the SDK is directly downloaded from the store, from
the CANDIDATE branch. This prevents building a RUNTIME snap with
a locally built SDK snap. But this is a very desirable thing to
do because it's the only way of testing a change in the SDK and
ensure that it doesn't break any snap that depends on them.

This patch includes a python script that modifies the snapcraft
file to build it using either a snap called gnome-46-2404-sdk.snap
placed in the project root, or the snap passed as its first
parameter.

This allows to build the RUNTIME locally with a new, still not
uploaded, SDK, and test it.

Update README.md

Use a script for local build

To make it safer, an external script is proposed, that modifies
the snapcraft.yaml file, builds the runtime with the desired
SDK, and restores the snapcraft.yaml to its original contents.

Move the YAML file into snap/ folder

This way, the script just creates a new snapcraft.yaml file that
takes precedence, and deletes it at completion. This is safer,
because the snap isn't left in an inconsistent state.

Fix README.md

Simpler way

Fix build

more fixes

extra change